Fix product catalogue: direct DB queries, URL-safe slugs, consistent caching
