+++ src/app/product/[id]/page.tsx
@@ -1,12 +1,9 @@
-// Read page: cached/revalidated hourly (consistent with the product API).
+// Read page: cached/revalidated hourly (consistent with the other read pages).
 export const revalidate = 60;
 
 import ProductDetail from '@/components/ProductDetails/v2/ProductDetails';
 import { notFound } from 'next/navigation';
-import urlMap from '@/lib/endpoint';
-import type { Product } from '@/lib/types';
-
-type ProductWithRelated = Product & { relatedProducts: Array<Product> };
+import { getProductById } from '@/lib/products';
 
 interface ProductPageProps {
   params: {
@@ -14,22 +11,8 @@ interface ProductPageProps {
   };
 }
 
-/**
- * Returns the product, or null when it doesn't exist (404).
- * Throws only on genuine server/network errors so those still surface.
- */
-const fetchProductById = async (id: string): Promise<ProductWithRelated | null> => {
-  const res = await fetch(urlMap.getProduct(id), { next: { revalidate: 60 } });
-
-  if (res.status === 404) return null;
-  if (!res.ok) {
-    throw new Error(`Failed to fetch product ${id}: ${res.status}`);
-  }
-  return res.json();
-};
-
 export async function generateMetadata({ params }: ProductPageProps) {
-  const product = await fetchProductById(params.id);
+  const product = await getProductById(params.id);
 
   if (!product) {
     return { title: 'Product Not Found | Syntara' };
@@ -42,7 +25,7 @@ export async function generateMetadata({ params }: ProductPageProps) {
 }
 
 export default async function ProductPage({ params }: ProductPageProps) {
-  const product = await fetchProductById(params.id);
+  const product = await getProductById(params.id);
 
   if (!product) {
     notFound();
